% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{collection_list}
\alias{collection_list}
\title{Search for a collection}
\usage{
collection_list(name = NULL, limit = 0, offset = 0)
}
\arguments{
\item{name}{If given, search for collections containing this word in the title}

\item{limit}{If given, maximum number of sources to display}

\item{offset}{If given, offset for source list}
}
\value{
A tibble containing one collection per row, including an \code{id}
}
\description{
Search for a collection
}
\examples{
collection_list("college")
}
