% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{do_request}
\alias{do_request}
\title{Helper function to do an actual request to a specific endpoint}
\usage{
do_request(endpoint, params = NULL, api = "https://search.mediacloud.org/api/")
}
\arguments{
\item{endpoint}{a url path}

\item{params}{a named list of query parameters}
}
\description{
Helper function to do an actual request to a specific endpoint
}
