% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{normalize_params}
\alias{normalize_params}
\title{Helper function to check and normalize parameters}
\usage{
normalize_params(
  query,
  start_date,
  end_date,
  collection_ids = NULL,
  source_ids = NULL
)
}
\description{
Helper function to check and normalize parameters
}
