% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{source_list}
\alias{source_list}
\title{Search for a source}
\usage{
source_list(name = NULL, collection_id = NULL, limit = 0, offset = 0)
}
\arguments{
\item{name}{If given, search for source containing this word in the title}

\item{collection_id}{If given, search for sources within this collection}

\item{limit}{If given, maximum number of sources to display}

\item{offset}{If given, offset for source list}
}
\value{
A tibble containing one collection per row, including an \code{id}
}
\description{
Search for a source
}
\examples{
collection_list("college")
}
