% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{story_count}
\alias{story_count}
\title{Get the story count for a specific query}
\usage{
story_count(
  query,
  start_date,
  end_date,
  collection_ids = NULL,
  source_ids = NULL
)
}
\arguments{
\item{query}{A query string}

\item{start_date}{The start date, should be a Date object}

\item{end_date}{The end date, should be a Date object}

\item{collection_ids}{Optional collection IDs}

\item{source_ids}{Optional source IDs}
}
\value{
the number of stories
}
\description{
Get the story count for a specific query
}
\examples{
story_count(query="jetten", start_date=as.Date("2024-01-01"), end_date=as.Date("2024-02-01"))
}
