% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{story_list}
\alias{story_list}
\title{Get the list of stories matching a query}
\usage{
story_list(
  query,
  start_date,
  end_date,
  collection_ids = NULL,
  source_ids = NULL,
  max_pages = 1
)
}
\arguments{
\item{query}{A query string}

\item{start_date}{The start date, should be a Date object}

\item{end_date}{The end date, should be a Date object}

\item{collection_ids}{Optional collection IDs}

\item{source_ids}{Optional source IDs}

\item{max_pages}{Maximum number of pages to retrieve (set to Inf to retrieve all)}
}
\value{
a tibble with one story per row
}
\description{
Get the list of stories matching a query
}
\examples{
story_list(query="jetten", start_date=as.Date("2024-01-01"), end_date=as.Date("2024-06-01"), max_pages=10)
}
